docshell/test/test_pushState_after_document_open.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=957479
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 957479</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 957479 **/
  SimpleTest.waitForExplicitFinish();
  // Child needs to invoke us, otherwise our onload will fire before the child
  // has done the write/close bit.
  onmessage = function doTest() {
    is(frames[0].location.pathname, "/tests/docshell/test/file_pushState_after_document_open.html", "Should have the right path here");
    is(frames[0].location.hash, "", "Should have the right hash here");
    frames[0].history.pushState({}, '', frames[0].document.URL + "#foopy");
    is(frames[0].location.pathname, "/tests/docshell/test/file_pushState_after_document_open.html", "Pathname should not have changed");
    is(frames[0].location.hash, "#foopy", "Hash should have changed");
    SimpleTest.finish();
  }

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=957479">Mozilla Bug 957479</a>
<p id="display"></p>
<div id="content" style="display: none">
<iframe src="file_pushState_after_document_open.html"></iframe>
</div>
<pre id="test">
</pre>
</body>
</html>
